requirements.txt
opencv-python
opencv-contrib-python
mysql-connector-python
streamlit
pandas
numpy
pillow
